Monaco editors and related cleaned and working. Just need a layer of paint to make it UX friendly
